Implement Diffusers backend and move optimization UI
