+++ api/backend/__init__.py
@@ -1,2 +1,1 @@
-from .models import *
 from .backend import *
