+++ api/models/prompt.py
@@ -1,0 +1,6 @@
+from dataclasses import dataclass
+
+@dataclass
+class Prompt:
+    positive: str
+    negative: str | None
