+++ generator_process/models/__init__.py
@@ -1,1 +1,5 @@
 from .fix_it_error import *
+from .image_generation_result import *
+from .optimizations import *
+from .scheduler import *
+from .upscale_tiler import *
